implement modular endpoints
